requirements.txt
# FastAPI - Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# Voice Activity Detection
webrtcvad==2.0.10

# Audio processing
pydub==0.25.1

# Speech-to-text (Whisper)
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Arabic text processing
pyarabic==0.6.15

# Database
psycopg2-binary==2.9.9

# Utilities
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# Additional dependencies
ffmpeg-python==0.2.0